makes useSelect hook
